@@ -1,39 +1,33 @@
 {
-	"name": "peerbox",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"dev": "webpack serve --mode development",
-		"build": "webpack --mode production"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "MIT",
-	"devDependencies": {
-		"@babel/core": "^7.19.3",
-		"@babel/preset-env": "^7.19.3",
-		"@babel/preset-react": "^7.18.6",
-		"autoprefixer": "^10.4.12",
-		"babel-loader": "^8.2.5",
-		"copy-webpack-plugin": "^11.0.0",
-		"css-loader": "^6.7.1",
-		"css-minimizer-webpack-plugin": "^4.2.0",
-		"html-webpack-plugin": "^5.5.0",
-		"mini-css-extract-plugin": "^2.6.1",
-		"postcss": "^8.4.17",
-		"postcss-loader": "^7.0.1",
-		"style-loader": "^3.3.1",
-		"tailwindcss": "^3.1.8",
-		"webpack": "^5.74.0",
-		"webpack-cli": "^4.10.0",
-		"webpack-dev-server": "^4.11.1",
-		"webpack-merge": "^5.8.0"
-	},
-	"dependencies": {
-		"nanoid": "^4.0.0",
-		"peerjs": "^1.4.7",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0"
-	}
+  "name": "peerbox",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "autoprefixer": "^10.4.20",
+    "nanoid": "^5.0.9",
+    "peerjs": "^1.5.4",
+    "postcss-import": "^16.1.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.15"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.15.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^9.15.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.14",
+    "globals": "^15.12.0",
+    "vite": "^6.0.1"
+  }
 }
