@@ -1,28 +1,26 @@
-/** @type {import('tailwindcss').Config} */
-const plugin = require("tailwindcss/plugin");
+import plugin from "tailwindcss/plugin";
 
-module.exports = {
-  content: ["./src/**/*.{html,js,jsx}","./public/index.html"],
+export default {
+  content: ["./src/**/*.{html,js,jsx}", "./index.html"],
   theme: {
     extend: {},
   },
   plugins: [
     plugin(function ({ addUtilities }) {
       addUtilities({
-        '.no-scrollbar': {
+        ".no-scrollbar": {
           /* IE and Edge */
-          '-ms-overflow-style': 'none',
+          "-ms-overflow-style": "none",
 
           /* Firefox */
-          'scrollbar-width': 'none',
+          "scrollbar-width": "none",
 
           /* Safari and Chrome */
-          '&::-webkit-scrollbar': {
-            display: 'none'
-          }
-        }
-      }
-      )
-    })
+          "&::-webkit-scrollbar": {
+            display: "none",
+          },
+        },
+      });
+    }),
   ],
-}
+};
